revert temp code and cleanup
