+++ libs/react/components/src/lib/Sidebar/SidebarTheme/index.ts
@@ -1,0 +1,2 @@
+export { SidebarThemeProvider, useSidebarTheme } from './context';
+export type { SidebarTheme, SidebarVariant } from './types';
