+++ libs/react/shelter-operator/src/lib/components/AppSidebar/AppSidebar.tsx
@@ -2,8 +2,13 @@ import { Divider, Sidebar } from '@monorepo/react/components';
 import { mergeCss } from '@monorepo/react/shared';
 import { operatorPath } from '@monorepo/react/shelter';
 import { useState } from 'react';
-import { useLocation } from 'react-router-dom';
-import { isShelterManageRoute, paths, shelterManageRoute } from '../../routing';
+import { useLocation, useParams } from 'react-router-dom';
+import {
+  isShelterManageRoute,
+  isShelterRoute,
+  paths,
+  shelterManageRoute,
+} from '../../routing';
 import { ShelterProfileLinks } from './ShelterProfileLinks';
 
 type IProps = {
@@ -17,9 +22,7 @@ export function AppSidebar(props: IProps) {
 
   const [isOpen, setIsOpen] = useState(initialOpenState);
   const location = useLocation();
-  // const { shelterId } = useParams<{ shelterId: string }>();
-
-  const shelterId = '5';
+  const { shelterId } = useParams<{ shelterId: string }>();
 
   const parentCss = ['bg-[#FAFAFA]', className];
 
@@ -51,27 +54,29 @@ export function AppSidebar(props: IProps) {
           Users
         </Sidebar.Link>
 
-        <>
-          <Divider
-            className="my-4 h-6 text-[#747A82]"
-            lineClassName="bg-[#747A82]"
-            label={isOpen ? 'shelter management' : ''}
-          />
+        {isShelterRoute(location.pathname) && shelterId && (
+          <>
+            <Divider
+              className="my-4 h-6 text-[#747A82]"
+              lineClassName="bg-[#747A82]"
+              label={isOpen ? 'shelter management' : ''}
+            />
 
-          <Sidebar.Link
-            to={shelterManageRoute(shelterId)}
-            isActive={isShelterManageRoute(location.pathname)}
-            collapsed={!isOpen}
-          >
-            Operations
-          </Sidebar.Link>
+            <Sidebar.Link
+              to={shelterManageRoute(shelterId)}
+              isActive={isShelterManageRoute(location.pathname)}
+              collapsed={!isOpen}
+            >
+              Operations
+            </Sidebar.Link>
 
-          <ShelterProfileLinks
-            pathname={location.pathname}
-            shelterId={shelterId}
-            isOpen={isOpen}
-          />
-        </>
+            <ShelterProfileLinks
+              pathname={location.pathname}
+              shelterId={shelterId}
+              isOpen={isOpen}
+            />
+          </>
+        )}
       </Sidebar.Content>
     </Sidebar>
   );
